Add Task-Based Parallel Execution for Logs Fetching
	•	Implemented BaseTask to encapsulate task logic, making it reusable and modular for parallel execution in isolates.
	•	Created FetchLogsTask to fetch logs with offset and limit, processing file streams in a separate isolate.
	•	Integrated FetchLogsTask in LogsRepository, enabling parallel log fetching.
	•	Improved repository execution using the parallelism executor.
	•	Addressed limitations of isolate execution, such as logger unavailability.
	•	Prepared the architecture for better handling of tasks with specific parameters.
