+++ lib/features/settings/domain/repository/i_logs_repository.dart
@@ -1,6 +1,5 @@
 import 'package:hommie/features/settings/domain/entities/logs.dart';
 
 abstract class ILogsRepository {
-  Stream<Log> streamLogs();
   Future<List<Log>> fetchLogs(int offset, int limit);
 }
